+++ NEWS.md
@@ -1,3 +1,20 @@
+# jsTreeR 1.0.0 (2020-09-22)
+
+* Better Shiny example 'SuperTinyIcons'.
+* Removed `jsonlite` dependency.
+* Allow to pass a list of options to the `contextMenu` argument.
+* The Shiny value of the tree is now updated when the user 
+creates/deletes/renames/pastes a node.
+* New Shiny value `jsTreeMoved` triggered when a node is moved.
+* New Shiny value `jsTreeRenamed` triggered when a node is renamed.
+* New Shiny value `jsTreeCopied` triggered when a node is copied.
+* New Shiny value `jsTreeDeleted` triggered when a node is deleted.
+* New Shiny value `jsTreeCreated` triggered when a node is created.
+* New function `folderGadget`, which launches a Shiny gadget allowing to 
+manipulate one or more folders.
+
+
+
 # jsTreeR 0.1.0 (2020-08-26)
 
 First release.